fix jui error
